fix: satisfy modularity guard for commission ownership
