+++ packages/domain-membership-billing/src/enterprise-controls.test.ts
@@ -146,27 +146,6 @@ describe('finance control', () => {
     expect(result.ok).toBe(true);
   });
 
-  it('keeps agent commissions payable after read-scope bindings are revoked', () => {
-    const result = assertFinancePayability({
-      commissionId: 'comm-read-only-attribution',
-      subscriptionAgentId: 'agent-1',
-      userAgentId: 'agent-1',
-      agentClientAgentIds: [],
-      subscription: { status: 'active', cancelAtPeriodEnd: false, gracePeriodEndsAt: null },
-    });
-
-    expect(result).toEqual({
-      ok: true,
-      value: expect.objectContaining({
-        ownership: expect.objectContaining({
-          agentId: 'agent-1',
-          ownerType: 'agent',
-          resolvedFrom: 'subscription.agentId',
-        }),
-      }),
-    });
-  });
-
   it('composes ownership and lifecycle violations into one finance result', () => {
     const result = assertFinancePayability({
       commissionId: 'comm-blocked',
